robust loop closure with icp
fixed a bug of g2o edge direction

switched to initial hessian if dso fails

added test on KAIST Urban dataset

implmented SC ring key to speed up query

combined SC and ringkey extraction

+++ src/loop_closure/place_recognition/process_pts/pts_preprocess.h
@@ -19,7 +19,7 @@ inline void
 get_pts_spherical(const IDPose *cur_pose,
                   std::vector<IDPtIntensity *> &pts_nearby,
                   std::vector<std::pair<Eigen::Vector3d, float>> &pts_spherical,
-                  double lidarRange, double voxelAngle) {
+                  float lidarRange, float voxelAngle) {
   std::vector<IDPtIntensity *> new_pts_nearby, pts_spherical_raw;
   for (auto &p : pts_nearby) {
     Eigen::Vector4d p_g(p->pt(0), p->pt(1), p->pt(2), 1.0);
